Fix Unity compile errors for HUD/RS selection meshes and match seed.
Adds missing SelectionRect/Triangle mesh helpers and aligns MatchAnalytics seed typing so Asterra.Gameplay builds again.

+++ Assets/Asterra/Gameplay/MatchBootstrap.cs
@@ -772,7 +772,7 @@ namespace Asterra.Gameplay
             IsMatchRunning = true;
             Debug.Log($"[Asterra] Match started mode={playMode} seed={matchSeed} map={MapKey} players={_participants.Count}");
             CloudBuildInfo.LogBootstrap();
-            MatchAnalytics.RecordMatchStart(MapKey, playMode, _participants.Count, matchSeed);
+            MatchAnalytics.RecordMatchStart(MapKey, playMode, _participants.Count, (int)matchSeed);
         }
 
         private void BindTerrainTexturePaint()
